Upgrade akka 2.4.17, http 10.0.4, cassandra-query 0.23
Move to EventsByTagQuery2, since (1) is deprecated, and akka 2.5 will
promote v2 to be the actual EventsByTagQuery. Plus, the cassandra journal
now only supports time-based UUIDs as offset for queries (not just
unix timestamps).

For the moment, we keep our own protobuf EventEnvelope the same, as we
don't use its timestamp for anything authoritative. The field is renamed
though, to make space for possibly introducing TimeBasedUUID in the future.