Make command handlers asynchronous-first
When implementing actors, we are seeing more and more async "setup" code
needing to do checks, pull in extra information, etc, taking place before
the real Handler is invoked. This asyncronous activity should ideally
be part of the Handler class itself.

This commit re-defines AbstractCommandHandler as a partial function from
(Command, State) to CompletionStage(Results), where Results encapsulates
what was previously in AbstractCommandHandler.

Existing implementations need to rework their code, with examples present
in the changes to ReplicatedTestActor, NonReplicatedTestActor and
AbstractStatefulPersistentActorSpec in this commit.