Fail/stop persistent actor if a handler fails asynchronously
Since we use pipe() to receive the results of the handler's
CompletionStage, we may be sent a Failure message if the completion
stage fails with an exception. This used to just go unhandled.

A good default is to re-throw, letting the actor's supervisor deal with
the failure. Implementations can still provide a custom createReceive()
to override what's done with Failure.