Allow createReceiveRecover to be invoked multiple times
Classes extending ReplicatedActor may want to refer to its
createReceiveRecover() method multiple times. Currently, this causes the
base class to not work, since its AtomicReference would be created
several times.

By assigning the Receive to a variable, and having the method just
return that, we can invoke it as often as we like (like a scala val).