Update kamon stats for multiple workers, and implement reset
The statistic type for the time-based worker stats was wrong: since it's
a slow-changing value with only 1 global value, all of those should be
"Gauge" type and not "Histogram". An additional problem is that
histograms, by default, are limited to 2.1E12 as maximum value, which is
no good for time stamps.

We also reimplement "Reset" functionality, so we can easily trigger a
complete reimport; but now, we retain any extra workers that are already
present.