Add a hook to inform the subclasses about materialization progress
In MaterializerActor, the subclasses can choose to materialize events
in parallel by choosing a concurrency key. While this is useful from
performance perspective, the importer looses the ability to track the
overall progress of the materialization.
This commit adds a method which is called whenever the materialization
workers report progress. The subclasses can override this method to
keep track of the materializartion progress.