Enable addition of custom tags to MaterializerActor metrics
The only base tag attached to all MaterializerActor metrics is the class
name of the actor. This can lead to duplicate metric names
if a single subclass is instantiated multiple times to do different
materializations. To remedy this and also facilitate further adjustments
of metric tags, this commit adds a protected method so that the
subclasses can introduce new tags to the metrics.