Skip the simple variant of replication test
The replication test currently is very unreliable, because it actually
starts 4 separate cassandra instances. Until we have time to simply run
it against one (in different keyspaces for each ActorSystem), let's
disable the "simple" variant of the test. That way, we at least cut it
down to only 2 cassandra instances.