Properly end a worker if not all timestamps are in the stream
We are seeing workers not always being ended, and that may be for two
reasons:
1. The current MaterializerActor never sends events with t=endTimestamp
to MaterializerWorkers, yet the latter requires t=endTimestamp to
conclude the worker is done.
2. The exact endTimestamp might not even occur in the source stream, if
it's sparse enough.

So, instead, we should wait until we've progressed _past_ the
endTimestamp, finish that exact timestamp, and then MaterializerWorkers
should conclude that we are done.